fix: Complete remaining TODOs and unify DocId types
Address all remaining TODO comments in source code and resolve
type inconsistency between index and storage layers.

TODO implementations:
- Extract filters from binlog row data automatically
- Implement SHOW BINARY LOG STATUS for latest GTID (with fallback to SHOW MASTER STATUS)
- Add state file GTID reading support
- Apply filter conditions in SEARCH and COUNT commands
- Update MASTER/SLAVE terminology to PRIMARY/REPLICA with legacy note

Type system improvements:
- Unify DocId type to uint32_t across index and storage layers
- Supports up to 4B documents (sufficient for practical use)
- Maintains memory efficiency (vs uint64_t which doubles memory usage)
- Compatible with Roaring Bitmap native support

Filter implementation:
- Add filter extraction in WRITE_ROWS, UPDATE_ROWS, DELETE_ROWS events
- Implement filter matching in tcp_server for SEARCH/COUNT queries
- Support equality operator with type-safe comparison using std::visit

Build system:
- Add .gitignore entries for test config files and SQL dumps
- Include storage/gtid_state.h in main.cpp for GTIDStateFile

All 163 tests passing.

+++ src/storage/document_store.h
@@ -15,7 +15,7 @@
 namespace mygramdb {
 namespace storage {
 
-using DocId = uint64_t;
+using DocId = uint32_t;  // Supports up to 4B documents (aligned with index::DocId)
 
 /**
  * @brief Filter value types
